Move labels_by_id methods to service
