+++ breadbox/breadbox/crud/dataset.py
@@ -59,6 +59,7 @@ from breadbox.io.filestore_crud import (
 )
 from .metadata import cast_tabular_cell_value_type
 from .dataset_reference import add_id_mapping
+from breadbox.service import metadata as metadata_service
 import typing
 
 log = logging.getLogger(__name__)
@@ -1038,41 +1039,6 @@ def get_tabular_dataset_index_given_ids(
     return [cell.dimension_given_id for cell in cells_in_id_column]
 
 
-def get_dataset_feature_labels_by_id(
-    db: SessionWithUser, user: str, dataset: Dataset,
-) -> dict[str, str]:
-    """
-    Try loading feature labels from metadata.
-    If there are no labels in the metadata or there is no metadata, then just return the feature names.
-    """
-    metadata_labels_by_given_id = get_dataset_feature_annotations(
-        db=db, user=user, dataset=dataset, metadata_col_name="label"
-    )
-
-    if metadata_labels_by_given_id:
-        return metadata_labels_by_given_id
-    else:
-        all_dataset_features = get_dataset_features(db=db, dataset=dataset, user=user)
-        return {feature.given_id: feature.given_id for feature in all_dataset_features}
-
-
-def get_dataset_sample_labels_by_id(
-    db: SessionWithUser, user: str, dataset: Dataset,
-) -> dict[str, str]:
-    """
-    Try loading sample labels from metadata.
-    If there are no labels in the metadata or there is no metadata, then just return the sample names.
-    """
-    metadata_labels = get_dataset_sample_annotations(
-        db=db, user=user, dataset=dataset, metadata_col_name="label"
-    )
-    if metadata_labels:
-        return metadata_labels
-    else:
-        samples = get_dataset_samples(db=db, dataset=dataset, user=user)
-        return {sample.given_id: sample.given_id for sample in samples}
-
-
 # TODO: This can probably be merged.
 def get_dataset_feature_annotations(
     db: SessionWithUser, user: str, dataset: Dataset, metadata_col_name: str,
@@ -1478,7 +1444,9 @@ def get_dataset_feature_by_label(
     assert_user_has_access_to_dataset(dataset, db.user)
     assert isinstance(dataset, MatrixDataset)
 
-    labels_by_given_id = get_dataset_feature_labels_by_id(db, db.user, dataset)
+    labels_by_given_id = metadata_service.get_dataset_feature_labels_by_id(
+        db, db.user, dataset
+    )
     given_ids_by_label = {label: id for id, label in labels_by_given_id.items()}
     feature_given_id = given_ids_by_label.get(feature_label)
     if feature_given_id is None:
@@ -1500,7 +1468,9 @@ def get_dataset_sample_by_label(
     assert_user_has_access_to_dataset(dataset, db.user)
     assert isinstance(dataset, MatrixDataset)
 
-    labels_by_given_id = get_dataset_sample_labels_by_id(db, db.user, dataset)
+    labels_by_given_id = metadata_service.get_dataset_sample_labels_by_id(
+        db, db.user, dataset
+    )
     given_ids_by_label = {label: id for id, label in labels_by_given_id.items()}
     sample_given_id = given_ids_by_label.get(sample_label)
     if sample_given_id is None:
@@ -1766,11 +1736,15 @@ def get_subsetted_matrix_dataset_df(
 
     # Re-index by label if applicable
     if dimensions_info.feature_identifier == FeatureSampleIdentifier.label:
-        labels_by_id = get_dataset_feature_labels_by_id(db, user, dataset)
+        labels_by_id = metadata_service.get_dataset_feature_labels_by_id(
+            db, user, dataset
+        )
         df = df.rename(columns=labels_by_id)
 
     if dimensions_info.sample_identifier == FeatureSampleIdentifier.label:
-        label_by_id = get_dataset_sample_labels_by_id(db, user, dataset)
+        label_by_id = metadata_service.get_dataset_sample_labels_by_id(
+            db, user, dataset
+        )
         df = df.rename(index=label_by_id)
 
     return df
